notion-gemini-embeddings/requirements.txt
fastapi
google-genai
httpx
notion-client
pgvector
psycopg2-binary
pydantic
pydantic-settings
pytest
PyYAML
python-dotenv
python-json-logger
tenacity
uvicorn
numpy
pandas
plotly
umap-learn
